// src/api/production-history/models/MediaAssetResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const mediaAssetResponseTypeEnum = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
} as const;

export type MediaAssetResponseTypeEnumKey =
  (typeof mediaAssetResponseTypeEnum)[keyof typeof mediaAssetResponseTypeEnum];

export type MediaAssetResponse = {
  /**
   * @type string
   */
  path: string;
  /**
   * @type string
   */
  physicalName: string;
  /**
   * @type string | undefined
   */
  originalName?: string;
  /**
   * @type string | undefined
   */
  type?: MediaAssetResponseTypeEnumKey;
  /**
   * @type string | undefined
   */
  mimeType?: string;
};