// src/api/production-history/models/PaginatedReportsResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ProductionHistoryRecordResponse } from './ProductionHistoryRecordResponse';

export type PaginatedReportsResponse = {
  /**
   * @type array
   */
  items: ProductionHistoryRecordResponse[];
  /**
   * @type number
   */
  total: number;
  /**
   * @type number
   */
  page: number;
  /**
   * @type number
   */
  limit: number;
  /**
   * @type number
   */
  pages: number;
};