// src/api/production-history/models/ProductContextResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ProductContextResponse = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  orderNum: string;
  /**
   * @type number
   */
  productionId: number;
  /**
   * @type string
   */
  partCode: string;
  /**
   * @type string | undefined
   */
  partName?: string;
  /**
   * @type string
   */
  serialNumber: string;
  /**
   * @type string | undefined
   */
  serverTime?: string;
};