// src/api/production-history/models/ProductionHistoryRecordResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { MediaAssetResponse } from './MediaAssetResponse';

export const productionHistoryRecordResponseFeatureColumnEnum = {
  MONTAGEM: 'MONTAGEM',
  QUALIDADE: 'QUALIDADE',
  INSPECAO: 'INSPECAO',
} as const;

export type ProductionHistoryRecordResponseFeatureColumnEnumKey =
  (typeof productionHistoryRecordResponseFeatureColumnEnum)[keyof typeof productionHistoryRecordResponseFeatureColumnEnum];

export const productionHistoryRecordResponseStatusEnum = {
  UPLOADED: 'UPLOADED',
  FAILED: 'FAILED',
} as const;

export type ProductionHistoryRecordResponseStatusEnumKey =
  (typeof productionHistoryRecordResponseStatusEnum)[keyof typeof productionHistoryRecordResponseStatusEnum];

export type ProductionHistoryRecordResponse = {
  /**
   * @type string
   */
  _id: string;
  /**
   * @type string | undefined
   */
  code?: string;
  /**
   * @type string | undefined
   */
  appName?: string;
  /**
   * @type string | undefined
   */
  pn?: string;
  /**
   * @type string | undefined
   */
  serialNumber?: string;
  /**
   * @type string | undefined
   */
  productionId?: string;
  /**
   * @type string | undefined
   */
  codItem?: string;
  /**
   * @type string | undefined
   */
  gate?: string;
  /**
   * @type string
   */
  featureColumn: ProductionHistoryRecordResponseFeatureColumnEnumKey;
  /**
   * @type string
   */
  status: ProductionHistoryRecordResponseStatusEnumKey;
  /**
   * @type number | undefined
   */
  attempts?: number;
  /**
   * @type array
   */
  media: MediaAssetResponse[];
};